Add shader extraction tool for rfg (.fxo_kg files)
Mostly ignores the rfg specific header at the beginning of the file, finds all the dx11 shaders inside the fxo_kg file, and extracts them, and disassembles them using a library called SlimShader. A lot of room for improvement here such as allowing editing and repacking, or writing a parser that makes the disassembly look more like real code.